added remove_from_cart api and made the apis restful

+++ cart/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, get_object_or_404
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
-from .serializers import CartItemSerializer, CartItemUpdateSerializer
+from .serializers import CartItemUpdateSerializer, AddToCartSerializer, CartItemDisplaySerializer
 from products.models import Product
 from .models import Cart, CartItem
 from rest_framework.response import Response
@@ -19,7 +19,7 @@ def cart_items_list_create(request):
 
       cart_items = CartItem.objects.filter(cart=cart)
 
-      serializer = CartItemSerializer(cart_items, many=True)
+      serializer = CartItemDisplaySerializer(cart_items, many=True)
 
       return Response({
         'data': serializer.data
@@ -34,7 +34,7 @@ def cart_items_list_create(request):
 
   if request.method == "POST":
 
-    serializer = CartItemSerializer(data=request.data)
+    serializer = AddToCartSerializer(data=request.data)
 
     if serializer.is_valid():
       product = serializer.validated_data.get('product')
@@ -75,7 +75,7 @@ def cart_items_list_create(request):
             'message': "cartItem created successfully",
             'data': {
               'cart': cart_item_obj.cart.id,
-              'product_id': cart_item_obj.id,
+              'product_id': cart_item_obj.product.id,
               'product': cart_item_obj.product.name,
               'quantity': cart_item_obj.quantity
             }
@@ -88,63 +88,62 @@ def cart_items_list_create(request):
 
 
 
-@api_view(['PATCH'])
+@api_view(['PATCH', 'DELETE'])
 @permission_classes([IsAuthenticated])
-def update_cart_item_quantity(request, cart_item_id):
+def cart_item_detail(request, cart_item_id):
 
-  cart_item = get_object_or_404(CartItem, id=cart_item_id)
+  if request.method == "PATCH":
 
-  if cart_item.cart.user != request.user:
+    cart_item = get_object_or_404(CartItem, id=cart_item_id)
+
+    if cart_item.cart.user != request.user:
+
+      return Response({
+        'error': "You don't have the permission."
+      }, status=status.HTTP_403_FORBIDDEN)
+    
+    serializer = CartItemUpdateSerializer(cart_item, data=request.data, partial=True)
+
+    if serializer.is_valid():
+
+      quantity = serializer.validated_data.get('quantity')
+
+      if quantity == 0:
+
+        cart_item.delete()
+
+        return Response({
+          'message': "the cartitem is removed from the cart"
+        })
+
+      if cart_item.product.stock_quantity >= quantity:
+
+        serializer.save()
+
+        return Response({
+          'message': "quantity updated successfully",
+          'data': serializer.data
+        }, status=status.HTTP_200_OK)
+      else:
+        return Response({
+          'error': "out of stock."
+        }, status=status.HTTP_400_BAD_REQUEST)
+    
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+  
+
+
+  if request.method == "DELETE":
+    cart_item = get_object_or_404(CartItem, id=cart_item_id)
+
+    if cart_item.cart.user != request.user:
+
+      return Response({
+        'error': "you are not allowed to perform this operation."
+      }, status=status.HTTP_403_FORBIDDEN)
+    
+    cart_item.delete()
 
     return Response({
-      'error': "You don't have the permission."
-    }, status=status.HTTP_403_FORBIDDEN)
-  
-  serializer = CartItemUpdateSerializer(cart_item, data=request.data, partial=True)
-
-  if serializer.is_valid():
-
-    quantity = serializer.validated_data.get('quantity')
-
-    if quantity == 0:
-
-      cart_item.delete()
-
-      return Response({
-        'message': "the cartitem is removed from the cart"
-      })
-
-    if cart_item.product.stock_quantity >= quantity:
-
-      serializer.save()
-
-      return Response({
-        'message': "quantity updated successfully",
-        'data': serializer.data
-      }, status=status.HTTP_200_OK)
-    else:
-      return Response({
-        'error': "out of stock."
-      }, status=status.HTTP_400_BAD_REQUEST)
-  
-  return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-
-@api_view(["DELETE"])
-@permission_classes([IsAuthenticated])
-def remove_from_cart(request, cart_item_id):
-
-  cart_item = get_object_or_404(CartItem, id=cart_item_id)
-
-  if cart_item.cart.user != request.user:
-
-    return Response({
-      'error': "you are not allowed to perform this operation."
-    }, status=status.HTTP_403_FORBIDDEN)
-  
-  cart_item.delete()
-
-  return Response({
-    'message': "cart item is successfully removed from cart."
-  }, status=status.HTTP_200_OK)
+      'message': "cart item is successfully removed from cart."
+    }, status=status.HTTP_200_OK)
